Update to Inserting API Key
Newer updates to inserting the API key in just quotes have been removing the quotes BuildConfig.java causing errors. Since the user cannot save changes in the BuildConfig.java because it was generated, the API key must have the the escape character added to have the double quotes carry over to BuildConfig.java and run properly.